Rename `npm.js` to `index.js` #34 When importing the module, by default the `index.js` gets recognized and in some scenarios the `npm.js` will not be detected by default.
