+++ package.json
@@ -7,7 +7,7 @@
     "author": "Chris Veness",
     "version": "1.1.1",
     "license": "MIT",
-    "main": "npm.js",
+    "main": "index.js",
     "bugs": "https://github.com/chrisveness/geodesy/issues",
     "scripts": {
         "test": "mocha test/*.js",
