+++ test/os-gridref-tests.js
@@ -6,9 +6,9 @@
 
 var chai      = require('chai');  // BDD/TDD assertion library
 
-var LatLon    = require('../npm.js').LatLonEllipsoidal;
-var OsGridRef = require('../npm.js').OsGridRef;
-var Dms       = require('../npm.js').Dms;
+var LatLon    = require('../').LatLonEllipsoidal;
+var OsGridRef = require('../').OsGridRef;
+var Dms       = require('../').Dms;
 
 chai.should();
 var test = it; // just an alias
